refactor(user-context-menu): remove unused context menu components
Clean up the user context menu by removing unused and commented-out components to improve code readability and maintainability. This change focuses on simplifying the component and reducing unnecessary clutter.
