+++ app/[locale]/(main)/admin/setting/(users)/user-management.card.tsx
@@ -14,10 +14,14 @@ type Props = {
 
 export function UserManagementCard({ user }: Props) {
 	return (
-		<div className="grid w-full grid-cols-[1fr_10rem_auto] gap-4 bg-card px-4 py-2 rounded- items-center rounded-md overflow-hidden">
+		<div className="grid w-full grid-cols-[1fr_10rem_auto] gap-4 bg-card px-2 py-1 items-center rounded-lg overflow-hidden border">
 			<div className="flex justify-between space-x-2 overflow- items-center overflow-hidden">
 				<UserAvatar user={user} url />
-				<CopyButton className="w-full justify-start overflow-hidden hover:bg-transparent bg-transparent border-transparent" data={user.id} content={user.id}>
+				<CopyButton
+					className="w-full justify-start overflow-hidden hover:bg-transparent bg-transparent border-transparent"
+					data={user.id}
+					content={user.id}
+				>
 					<h3>{user.name}</h3>
 				</CopyButton>
 			</div>
