+++ app/[locale]/(main)/admin/setting/(users)/user.table.tsx
@@ -53,7 +53,7 @@ export function UserTable() {
 	return (
 		<div className="flex h-full w-full flex-col space-y-2 overflow-hidden">
 			<div>
-				<div className="flex h-14 gap-2 p-2 bg-card rounded-md">
+				<div className="flex h-10 gap-2">
 					<Input
 						className="h-full"
 						value={name}
